Module №5. WEB-dev/7. Django Rest Framework/pyproject.toml
[tool.poetry]
name = "django rest"
version = "0.1.0"
description = ""
authors = ["usik-sviatoslav <usik.sviatoslav.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
environs = "^9.5.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
